removed code from main loop to specific anaylis folder for cleaner code structure
